feat(settings): put the reading settings behind their own screen
The Reading section had grown to eight switches, a chip row and two
rows only some devices show, which is a list to read through rather
than a place to go. It moves behind one row, Settings → Reading &
navigation, the way how-the-page-looks already sits behind Reading
appearance.

The dictionary section goes with it: looking a word up is something
you do with a book open. The e-ink rows get a Screen header there,
since they describe the panel and not the reading.

The row says what it holds — "Read by scrolling" or "Paged.
Page-turn sides: Swapped" — and says nothing about sides in a
scrolled book, where there are none to tap.

SettingsGroup, RowDivider, SwitchRow and ChipRow move to
SettingsRows.kt so both screens draw the same rows; SettingsGroup's
help icon takes its description from the caller rather than naming
the library section it was written for.
